+++ src/main/java/com/dsmpear/main/payload/response/ProfileReportListResponse.java
@@ -12,7 +12,7 @@ public class ProfileReportListResponse extends PageResponse{
     private List<MyPageReportResponse> myPageReportResponses;
 
     @Builder
-    public ProfileReportListResponse(int totalElements, int totalPages, List<ProfileReportResponse> profileReportResponses, List<MyPageReportResponse> myProfileReportResponses) {
+    public ProfileReportListResponse(Long totalElements, int totalPages, List<ProfileReportResponse> profileReportResponses, List<MyPageReportResponse> myProfileReportResponses) {
         super(totalElements, totalPages);
         this.profileReportResponses = profileReportResponses;
         this.myPageReportResponses = myProfileReportResponses;
