+++ src/main/java/com/dsmpear/main/payload/response/ReportListResponse.java
@@ -13,7 +13,7 @@ public class ReportListResponse extends PageResponse {
     private List<ReportResponse> reportResponses;
 
     @Builder
-    public ReportListResponse(int totalElements, int totalPages, List<ReportResponse> reportResponses) {
+    public ReportListResponse(Long totalElements, int totalPages, List<ReportResponse> reportResponses) {
         super(totalElements, totalPages);
         this.reportResponses = reportResponses;
     }
